Record the exit status of the daemon when running as a service, and return it to Windows so that Windows doesn't tell the admin that "the service did not report an error" when it stopped unexpectedly.
When failing to contact the SCM, report a textual error message as
well as the error code.

Make OurService() take a const char * instead of char *, so that we can
pass it a std::string.c_str().

InstallService creates service using "-s" option instead of "--service",
which no longer works once we use getopt() for option processing
(to follow).