Move BackupStoreAccountsControl into same file as BackupAccountControl.
Rename it to BackupStoreAccountControl (without the "s") for consistency.

Change interface so that it's constructed with an account number, so it can
be omitted from each command, which helps us to maintain the same interface
for S3BackupAccountControl.

Refactor bbstoreaccounts to create space for Amazon S3 command implementation.

Fix confusion about URIs and paths in S3 filesystem.  Make it clear (with
parameter names and comments) which functions take paths (relative to the
configured base path), which functions take URIs (complete paths) and which
functions take URLs (hostname, port and URI).

Put all objects (files and directories) in the same "directory" on S3 so that
we can list them all together and easily identify the type of each object (from
its extension). Previously they were separated into "directories" (delimited
prefixes) called "files" and "dirs" (unlike BackupStore filesystems).