Fix test environment setup and execution for CMake on Unix.
Rename the test runner scripts to _t and _t-gdb universally for consistency.

Add a -c option to runtest.pl which uses a CMake environment instead of make
to build tests and prepare their environment, rather than just assuming CMake
on Windows and Make elsewhere.

Use cmake -E commands to copy and delete files and directories instead of
hard-coding Windows commands, so that it works on Unix too.