CMake Unibuild: change version numbers from cache variables to private
The advantage of allowing users to configure these was purely
hypothetical, versus the very real problems that occurred when we were
forced to upgrade a dependency because the version we wanted was no
longer available on the original website/URL, but the old version number
and hash would be stored in the CMake cache, which would then have to be
cleared for builds to start working again.

Upgrade ZLib to 1.2.11 and OpenSSL to 1.0.2j, so that Windows Unibuild works
again. Temporarily switch to SVN for PCRE to fix static libs issue.  Switch to
zlib .tar.gz for consistency and simplicity.